Added an Exit menu option
Added Exit menu option and fixed the balance to be a float instead of
long datatype.

+++ Bank Account/BankClass.h
@@ -4,7 +4,7 @@ class bank
 {
 public:
 	static int accnumber;
-	long balance;
+	float balance;
 	UserDetails d;
 	void getdata();
 	bank transfermoney(bank);
